добавляет routes
