+++ library/routes/index.js
@@ -1,0 +1,8 @@
+const express = require('express');
+const routes = express.Router();
+
+const booksRoutes = require('./books');
+
+routes.use(booksRoutes);
+
+module.exports = routes;
